this commit will add DTO

+++ src/main/java/com/miu/registration/Repository/StudentRepository.java
@@ -1,10 +1,9 @@
 package com.miu.registration.Repository;
 
-
 import com.miu.registration.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student,Integer> {
+public interface StudentRepository extends JpaRepository<Student,Long> {
 }
